MicroMax og DummyUciEngine compilerer altid i x86-mode lidt småjusteringer i indrykning og size_t

+++ Spil/MakeChessOpeningLib/MakeChessOpeningLib.cpp
@@ -10,7 +10,7 @@ int main(int argc, char **argv) {
     OpeningLibrary lib;
 
     DirList list = scandir(_T("*.chs"));
-    for(int i = 0; i < list.size(); i++) {
+    for(size_t i = 0; i < list.size(); i++) {
       DirListEntry &name = list[i];
       _tprintf(_T("Adding <%s>\n"), name.name);
       lib.addGame(name.name);
